services/job-ingestion-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
psycopg2-binary
requests
httpx
python-dotenv
apscheduler
